requirements.txt
pyserial
monitorcontrol 
